chor: code structure
